// src/main/components/safe-box/usecases/update-users/update-users-safe-box-request-dto.ts
export interface IUpdateUsersSafeBoxRequestDTO {
  usuarios_leitura: string[];
  usuarios_escrita: string[];
  tipo: string;
  usuarios_leitura_deletado: string[];
  usuarios_escrita_deletado: string[];
  criptografia: string;
  nome: string;
  anexos: string;
  descricao: string;
  conteudo: string;
  organizacao: string;
  id: string;
}